organize the code

+++ SocialMedia_API/Controllers/PostController.cs
@@ -47,50 +47,72 @@ namespace SocialMedia_API.Controllers
         public async Task<IActionResult> DeletePost(int id)
         {
             var post = await postRepository.GetById(id);
-            var notifications = await notificationRepository.GetAll();
-            foreach (var item in notifications)
+            if (post != null)
             {
-                 await notificationRepository.Delete(item);
-
+                var notifications = await notificationRepository.GetAll();
+                foreach (var item in notifications)  { await notificationRepository.Delete(item); }
+                await postRepository.Delete(post);
+                return Ok(post);
             }
-            await postRepository.Delete(post);
-            return Ok(post);
-            
+            return BadRequest();
         }
 
-
-        [HttpPost("Comment")]
-        public async Task<IActionResult> AddComment(Comment comment)
+        [HttpGet("GetById/{Id}")]
+        public async Task<IActionResult> GetPostById(int Id)
         {
-            comment.CreatedTime = DateTime.Now;
+            var post = await postService.GetPostDtoById(Id);
+            if (post != null)
+                return Ok(post);
+            
+            return BadRequest(); 
+           
+        }
 
-            Notification notification = new()
+        [HttpGet]
+        public async Task<IActionResult> Explore()
+        {
+            var Posts = await postService.GetExplore();
+            if (Posts != null)
             {
-                CreatedTime = DateTime.Now,
-                NotifierId = comment.UserId,
-                PostId = comment.PostId,
-                Type = NotificationType.Comment
-            };
-            notification.UserId = postRepository.GetById(comment.PostId).Result.UserId;
+                return Ok(Posts.OrderByDescending(n=>n.LikesCounter));
+            }
 
-            if (notification.UserId != notification.NotifierId)
-                await notificationRepository.Insert(notification);
+            return BadRequest(); 
+        }
 
-            await commentRepository.Insert(comment);
-            return Ok(comment);
-          
+        [HttpGet("GetFollowingPosts/{UserId}")]
+        public async Task<IActionResult> GetFollowingPosts(string UserId)
+        {
+            List<PostDTO> Posts = new();
+            var allPosts = await postRepository.GetAll();
+            var followers = await followRepository.GetAll();
+            var posts = allPosts
+                 .Where(p => followers
+                   .Where(f => f.FollowerId == UserId)
+                    .Select(f => f.FollowingId)
+                        .Contains(p.UserId))
+                            .ToList();
+
+            posts.AddRange(allPosts.Where(n => n.UserId == UserId));
+           
+            foreach (var item in posts)
+            {
+                var post = await postService.GetPostDtoById(item.Id);
+                Posts.Add(post);
+            }
+
+            return Ok(Posts.OrderByDescending(n=>n.Post.CreatedTime));
         }
 
         [HttpPost("Like")]
         public async Task<IActionResult> AddLike(Like like)
         {
-            var res = await likeRepository.GetAll();
-            var les = res.Any(n => n.UserId == like.UserId && n.PostId == like.PostId);
+            var allLikes = await likeRepository.GetAll();
+            var les = allLikes.Any(n => n.UserId == like.UserId && n.PostId == like.PostId);
             if (les == false)
             {
                 Notification notification = new()
                 {
-                    CreatedTime = DateTime.Now,
                     NotifierId = like.UserId,
                     PostId= like.PostId,
                     Type = NotificationType.Like   
@@ -99,117 +121,28 @@ namespace SocialMedia_API.Controllers
                 
                 if (notification.UserId != notification.NotifierId)
                     await notificationRepository.Insert(notification);
-                
-
-                
-                like.CreatedTime = DateTime.Now;
+                     
                 await likeRepository.Insert(like);
                 return Ok(true);
+
             }
-            var Liked = res.Where(n => n.UserId == like.UserId && n.PostId == like.PostId).FirstOrDefault();
+            var Liked = allLikes.Where(n => n.UserId == like.UserId && n.PostId == like.PostId).FirstOrDefault();
             await likeRepository.Delete(Liked);
+
             return Ok(false);
         }
 
         [HttpPost("Like/isLiked")]
         public async Task<IActionResult> IsLiked(Like like)
         {
-            var res = await likeRepository.GetAll();
-            var Liked = res.Where(n => n.UserId == like.UserId && n.PostId == like.PostId).FirstOrDefault();
+            var allLikes = await likeRepository.GetAll();
+            var Liked = allLikes.Where(n => n.UserId == like.UserId && n.PostId == like.PostId).FirstOrDefault();
             if(Liked != null)
             {
                  return Ok(true);
 
             }
             return Ok(false);
-        }
-        
-        [HttpGet("GetById/{Id}")]
-        public async Task<IActionResult> GetPostById(int Id)
-        {
-            var post = await postService.GetPostDtoById(Id);
-            if (post != null)
-                return Ok(post);
-            
-            else { return BadRequest(); }
-           
-        }
-
-
-
-        [HttpGet]
-        public async Task<IActionResult> Explore()
-        {
-            List<PostDTO> Posts = new();
-            var res = await postRepository.GetAll();
-            foreach (var item in res)
-            {
-                List<CommentDTO> Comments = new();
-                var user = await userManger.FindByIdAsync(item.UserId);
-                PostDTO post = new()
-                {
-                    Post = item,
-                    CommentsCounter =  commentRepository.GetAll().Result.Where(n=>n.PostId == item.Id).Count(),
-                    LikesCounter = likeRepository.GetAll().Result.Where(n => n.PostId == item.Id).Count(),
-                };
-                
-                var comments = commentRepository.GetAll().Result.Where(n => n.PostId == item.Id).ToList();
-                foreach (var comment in comments)
-                {
-                    var commentUser = await userManger.FindByIdAsync(comment.UserId);
-                    CommentDTO dtocomment = new();
-                    dtocomment.Comment = comment;
-                    if (commentUser != null)
-                    {
-                        dtocomment.UserFullName = $"{commentUser.FirstName} {commentUser.LastName}";
-                        dtocomment.UserImg = commentUser.Images;
-                    }
-
-                    Comments.Add(dtocomment);
-                }
-
-                post.Comments = Comments.OrderByDescending(n=>n.Comment.CreatedTime).ToList();
-
-
-                if (user != null)
-                {
-                    post.UserImg = user.Images;
-                    post.UserId = item.UserId;
-                    post.FullName = $"{user.FirstName} {user.LastName}";
-                }
-                
-                Posts.Add(post);
-            }
-            return Ok(Posts.OrderByDescending(n=>n.LikesCounter));
-        }
-
-
-        [HttpGet("GetFollowingPosts/{UserId}")]
-        public async Task<IActionResult> GetPosts(string UserId)
-        {
-
-            List<PostDTO> Posts = new();
-            var res = await postRepository.GetAll();
-            var followers = await followRepository.GetAll();
-
-            var posts = res
-                 .Where(p => followers
-                   .Where(f => f.FollowerId == UserId)
-                    .Select(f => f.FollowingId)
-                        .Contains(p.UserId))
-                            .ToList();
-
-            posts.AddRange(res.Where(n => n.UserId == UserId));
-           
-
-
-            foreach (var item in posts)
-            {
-                var post = await postService.GetPostDtoById(item.Id);
-                Posts.Add(post);
-            }
-
-            return Ok(Posts.OrderByDescending(n=>n.Post.CreatedTime));
         }
     }
 }
